% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.rest.R
\docType{package}
\name{arcgis.rest}
\alias{arcgis.rest}
\title{arcgis.rest: A package for querying spatial data from ArcGIS REST APIs}
\description{
The arcgis.rest package primarily revolves around a single function:
get_spatial_layer. This function will send a POST request to a URL and
return the appropriate spatial data as an sf object
}
\section{get_layer_by_<spatial> functions}{

The \code{get_layer_by_spatial} functions will perform a spatial query to
the ArcGIS REST API using the appropriate spatial type.
\code{get_layer_by_poly} will query spatial layers that lie the passed
polygon, for instance, and \code{get_layer_by_line} will return spatial
layers that intersect with a provided line. These defaults can be overridden
to change the behavior of how the spatial object is queried.
}

