% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_urls}
\alias{example_urls}
\alias{reykjanes_lava_flow_url}
\alias{wi_hydro_url}
\alias{wi_landcover_url}
\alias{wi_leaf_off_url}
\title{Various URLs used in examples}
\format{
Character strings of URLs

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
reykjanes_lava_flow_url

wi_hydro_url

wi_landcover_url

wi_leaf_off_url
}
\description{
These are URLs that are used for examples throughout the package
}
\keyword{datasets}
