% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_geometry_type}
\alias{get_geometry_type}
\title{Get Geometry Type}
\usage{
get_geometry_type(url)
}
\arguments{
\item{url}{A character string of a feature services URL}
}
\value{
A character string of the layers geometry type
}
\description{
This function will return the geometry type of layer housed on an ArcGIS REST
API server.
}
\examples{
base_wdnr_url <- "https://dnrmaps.wi.gov/arcgis/rest/services/"
hydro_path <- "WT_SWDV/WT_Inland_Water_Resources_WTM_Ext_v2/MapServer/3"
hydro_url <- paste0(base_wdnr_url, hydro_path)
get_geometry_type(url = hydro_url)
}
