% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_layers.R
\name{get_image_layer}
\alias{get_image_layer}
\title{Retrieve an image layer from an ArcGIS REST API}
\usage{
get_image_layer(
  url,
  sf_object = NULL,
  bbox = NULL,
  token = "",
  output_type = "png",
  ...
)
}
\arguments{
\item{url}{A character string of the url for the layer to pull}

\item{sf_object}{An \code{sf} object used for the bounding box}

\item{bbox}{Character string of the bounding box}

\item{token}{A character string of the token (if needed)}

\item{output_type}{Character. Should the returned layer be "raster" or "png"}

\item{...}{Additional arguments to pass to the ArcGIS REST API}
}
\value{
Either a "RasterLayer" object or a tibble of class "magick-image"
}
\description{
This function retrieves image layers present an ArcGIS
REST services API
}
\details{
This is a core function of the package. It retrieves image layers from
an ArcGIS REST API designated by the URL. These layers require a bounding
box to be queried, which is either taken from the \code{sf_object} argument
or directly from the \code{bbox} argument.

All of the querying parameters are sent via a POST request to the URL, so
if there are issues with passing additional parameters via \code{...}
first determine how they fit into the POST request and make adjustments as
needed. This syntax can be tricky if you're not used to it.
}
\examples{
mke_landuse <- get_image_layer(wi_landcover_url, mke_county)
}
