% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_layer_html}
\alias{get_layer_html}
\title{Pull the HTML body from a web page}
\usage{
get_layer_html(url)
}
\arguments{
\item{url}{Character. The URL of the web page}
}
\value{
A character string of the HTML body
}
\description{
Used internally to pull HTML for a layer's web page so that the call
doesn't have to be made twice in \code{\link{get_geometry_type}} if the
url provided there is for a raster layer.
}
