% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_layers.R
\name{get_layer_info}
\alias{get_layer_info}
\title{Retrieve metadata for a layer}
\usage{
get_layer_info(url, token = "")
}
\arguments{
\item{url}{A character string of the url for the layer to pull}

\item{token}{A character string of the token (if needed)}
}
\value{
A list of metadata fields
}
\description{
This function retrieves metadata for a layer.
}
\examples{
\dontrun{
# lava flows on Reykjanes (pr. 'rake-yah-ness') peninsula in Iceland
lava_flows_info <- get_layer_info(reykjanes_lava_flow_url)
}
}
