% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_layer_properties.R
\name{get_layer_legend}
\alias{get_layer_legend}
\title{Returns a legend for a raster layer}
\usage{
get_layer_legend(url)
}
\arguments{
\item{url}{A URL to a Map or Image Service layer}
}
\value{
A data.frame with two columns (color, values) and the number of
rows equal to the number of values in a layer
}
\description{
Raster layers are accompanied with legends to identify what the colors mean.
This function retrieves those legend values and returns them as a data.frame
with the associated RGB color values. This will likely be most useful for
plotting and analysis of map layers.
}
\examples{
\dontrun{
get_layer_legend(wi_landcover_url)
}
}
