% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_service_type}
\alias{get_service_type}
\title{Get Service Type of a REST endpoint}
\usage{
get_service_type(url, ...)
}
\arguments{
\item{url}{A character string of a valid layer URL}

\item{...}{Only used internally}
}
\value{
A character string defining the layer type
}
\description{
Get Service Type of a REST endpoint
}
\examples{
\dontrun{
get_service_type(reykjanes_lava_flow_url)
}
}
