% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layers.R
\name{plot_layer,RasterBrick-method}
\alias{plot_layer,RasterBrick-method}
\title{Plot a RasterBrick object}
\usage{
\S4method{plot_layer}{RasterBrick}(
  x,
  outline_poly = NULL,
  outline_size = 1.2,
  outline_color = "gray30",
  plot_pkg = "ggplot",
  ...
)
}
\arguments{
\item{RasterBrick}{}
}
\value{

}
\description{
Plot a RasterBrick object
}
\examples{
\dontrun{
wi_aerial <- get_map_layer(wi_leaf_off_url, wis_poly)
plot_layer(wi_aerial, outline_poly = wis_poly)
}
}
