% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layers.R
\name{plot_layer,RasterLayer-method}
\alias{plot_layer,RasterLayer-method}
\title{Plot a RasterLayer object}
\usage{
\S4method{plot_layer}{RasterLayer}(
  x,
  outline_poly = NULL,
  outline_size = 1.2,
  outline_color = "gray30",
  plot_pkg = "ggplot",
  ...
)
}
\arguments{
\item{RasterLayer}{}
}
\value{

}
\description{
Plot a RasterLayer object
}
\examples{
\dontrun{
wi_landcover <- get_map_layer(wi_landcover_url, wis_poly)
plot_layer(wi_landcover, outline_poly = wis_poly)
}
}
