% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layers.R
\name{plot_layer,RasterStack-method}
\alias{plot_layer,RasterStack-method}
\title{Plot a RasterStack object}
\usage{
\S4method{plot_layer}{RasterStack}(
  x,
  outline_poly = NULL,
  outline_size = 1.2,
  plot_pkg = "ggplot",
  legend = FALSE,
  ...
)
}
\arguments{
\item{RasterStack}{}
}
\value{

}
\description{
Plot a RasterStack object
}
\examples{
\dontrun{
wi_aerial <- get_map_layer(wi_leaf_off_url, wis_poly)
plot_layer(wi_aerial, outline_poly = wis_poly)
}
}
