% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layers.R
\name{plot_layer,sf-method}
\alias{plot_layer,sf-method}
\title{Plot an sf object}
\usage{
\S4method{plot_layer}{sf}(
  x,
  outline_poly = NULL,
  outline_size = 1.2,
  outline_color = "gray30",
  plot_pkg = "ggplot",
  ...
)
}
\arguments{
\item{sf}{}
}
\value{

}
\description{
Plot an sf object
}
\examples{
\dontrun{
plot_layer(wis_poly)
}
}
