% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raster.R
\name{raster_data,RasterLayer-method}
\alias{raster_data,RasterLayer-method}
\title{Convert RasterLayer into data.frame that can be used for plotting}
\usage{
\S4method{raster_data}{RasterLayer}(x)
}
\arguments{
\item{x}{A RasterLayer object}
}
\value{
A data.frame with 3 columns and \code{length(raster_object)} rows
}
\description{
This function is used internally by \code{\link{plot_layer}} to convert a
RasterLayer object to a data.frame that can be used for plotting with ggplot2
}
\examples{
\dontrun{
wi_landcover <- get_map_layer(wi_landcover_url, wis_poly)
wi_landcover_data <- raster_data(wi_landcover)
}
}
