% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raster.R
\name{raster_data,RasterStack-method}
\alias{raster_data,RasterStack-method}
\title{Convert RasterStack into data.frame that can be used for plotting}
\usage{
\S4method{raster_data}{RasterStack}(x)
}
\arguments{
\item{x}{A RasterStack object}
}
\value{
A data.frame with 3 columns and \code{length(raster_object)} rows
}
\description{
This function is used internally by \code{\link{plot_layer}} to convert a
RasterLayer object to a data.frame that can be used for plotting with ggplot2
}
\examples{
\dontrun{
wi_leaf_off_layer <- get_image_layer(wi_leaf_off_url, wis_poly)
wi_leaf_off_data <- raster_data(wi_leaf_off_layer)
}
}
