% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_example_raster}
\alias{sf_example_raster}
\alias{wi_landcover}
\alias{wi_aerial_imagery}
\title{Various example raster objects}
\format{
An object of class \code{RasterLayer} of dimension 400 x 400 x 1.

An object of class \code{RasterBrick} of dimension 200 x 200 x 3.
}
\source{
\href{https://dnrmaps.wi.gov/arcgis_image/rest/services}{Wisconsin DNR ArcGIS Image Server}
}
\usage{
wi_landcover

wi_aerial_imagery
}
\description{
These are raster objects that are used for examples throughout the package
}
\keyword{datasets}
