% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_rel_lookups}
\alias{sp_rel_lookups}
\alias{sp_rel_valid}
\alias{sp_rel_lookup}
\title{Tables used to lookup and explain which spatial relation types go with
different spatial queries}
\format{



}
\source{
sp_rel_valid -- \url{
  https://desktop.arcgis.com/en/arcmap/latest/extensions/
  data-reviewer/types-of-spatial-relationships-that-can-be-validated.htm
}

sp_rel_lookup -- \url{
  https://help.arcgis.com/en/webapi/wpf/apiref/
  ESRI.ArcGIS.Client~ESRI.ArcGIS.Client.Tasks.SpatialRelationship.html
}
}
\usage{
sp_rel_valid

sp_rel_lookup
}
\description{
sp_rel_lookup explains the various different types of spatial relationships
available through ArcGIS REST APIs. sp_rel_valid shows which spatial
relationships are valid with different geometry types being queried and
used to do spatial queries
}
\keyword{datasets}
