% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{valid_sp_rel}
\alias{valid_sp_rel}
\title{Check to see which spatial relation types are applicable to the feature
classes being queried and the sf objects use do to a spatial query}
\usage{
valid_sp_rel(fc1, fc2, pull = TRUE)
}
\arguments{
\item{fc1}{Character. The feature class type being queried. Available options
are "point", "multipoint", "line", or "area".}

\item{fc2}{Character. The geometry type of the sf object used to do a spatial
query. Available options are "point", "multipoint", "line", or "area".}

\item{pull}{Logical. Pull the available options (TRUE) or print all columns
of the sp_rel_valid data.frame for the appropriate fc1 and fc2}
}
\value{
Either a vector or filtered data.frame showing the appropriate
sp_rels for the given feature classes
}
\description{
Check to see which spatial relation types are applicable to the feature
classes being queried and the sf objects use do to a spatial query
}
\examples{
valid_sp_rel("line", "line")
}
