% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_geometry_type}
\alias{get_geometry_type}
\title{Get Geometry Type}
\usage{
get_geometry_type(url)
}
\arguments{
\item{url}{A character string of a feature services URL}
}
\value{
A character string of the layers geometry type
}
\description{
Get Geometry Type
}
\examples{
\dontrun{
get_geometry_type(reykjanes_lava_flow_url)
}
}
