% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_example_polys}
\alias{sf_example_polys}
\alias{iceland}
\alias{mke_county}
\alias{portage_county}
\alias{reykjanes}
\alias{wis_counties}
\alias{wis_poly}
\title{Various example sf polygons}
\format{
An object of class sf and data.frame:

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 72 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\source{
\code{\link[ggplot2]{map_data}}
}
\usage{
iceland

mke_county

portage_county

reykjanes

wis_counties

wis_poly
}
\description{
These are sf polygons that are used for examples throughout the package
}
\keyword{datasets}
