% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sf_objects}
\alias{sf_objects}
\alias{sf_line}
\alias{sf_point}
\alias{sf_points}
\alias{sf_polygon}
\title{Create sf objects from coordinates}
\usage{
sf_line(..., crs = 4326)

sf_point(..., crs = 4326)

sf_points(..., crs = 4326)

sf_polygon(..., crs = 4326)
}
\arguments{
\item{...}{The coordinates of the object}

\item{crs}{The coordinate reference system. Defaults to 4326}
}
\value{
An sf object of the appropriate type
}
\description{
These are simple wrapper functions for creating sf objects from points
}
\examples{
pt_a <- c(-90, 45)
pt_b <- c(-89, 44)
pt <- sf_points(pt_a)
line <- sf_line(pt_a, pt_b)
}
